[Refactor] query
