+++ src/main/java/com/alipay/oceanbase/rpc/stream/QueryResultSet.java
@@ -27,8 +27,6 @@ import java.util.Map;
 
 public class QueryResultSet {
     private final QueryStreamResult queryStreamResult;
-    private boolean                 hasMore;
-    private long                    sessionId;
 
     /*
      * Query result set.
@@ -78,21 +76,5 @@ public class QueryResultSet {
      */
     public void close() throws Exception {
         queryStreamResult.close();
-    }
-
-    public boolean isHasMore() {
-        return hasMore;
-    }
-
-    public void setHasMore(boolean hasMore) {
-        this.hasMore = hasMore;
-    }
-
-    public long getSessionId() {
-        return sessionId;
-    }
-
-    public void setSessionId(long sessionId) {
-        this.sessionId = sessionId;
     }
 }
