--- src/main/java/com/alipay/oceanbase/rpc/stream/async/ObTableQueryAsyncStreamResult.java
@@ -1,66 +1,0 @@
-/*-
- * #%L
- * OBKV Table Client Framework
- * %%
- * Copyright (C) 2021 OceanBase
- * %%
- * OBKV Table Client Framework is licensed under Mulan PSL v2.
- * You can use this software according to the terms and conditions of the Mulan PSL v2.
- * You may obtain a copy of Mulan PSL v2 at:
- *          http://license.coscl.org.cn/MulanPSL2
- * THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
- * EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
- * MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
- * See the Mulan PSL v2 for more details.
- * #L%
- */
-
-package com.alipay.oceanbase.rpc.stream.async;
-
-import com.alipay.oceanbase.rpc.location.model.partition.ObPair;
-import com.alipay.oceanbase.rpc.protocol.payload.ObPayload;
-import com.alipay.oceanbase.rpc.protocol.payload.impl.execute.query.AbstractQueryStreamResult;
-import com.alipay.oceanbase.rpc.protocol.payload.impl.execute.query.ObTableQueryResult;
-import com.alipay.oceanbase.rpc.protocol.payload.impl.execute.syncquery.ObTableQueryAsyncResult;
-import com.alipay.oceanbase.rpc.table.ObTable;
-import com.alipay.oceanbase.rpc.table.ObTableParam;
-
-public class ObTableQueryAsyncStreamResult extends AbstractQueryStreamResult {
-    private boolean isEnd = true;
-    private long    sessionId;
-
-    @Override
-    protected ObTableQueryResult execute(ObPair<Long, ObTableParam> partIdWithObTable,
-                                         ObPayload streamRequest) throws Exception {
-        throw new IllegalArgumentException("not support this execute");
-    }
-
-    @Override
-    protected ObTableQueryAsyncResult executeAsync(ObPair<Long, ObTableParam> partIdWithObTable,
-                                                   ObPayload streamRequest) throws Exception {
-        Object result = partIdWithObTable.getRight().getObTable().execute(streamRequest);//执行query start/ query next等等
-
-        cacheStreamNext(partIdWithObTable, checkObTableQueryAsyncResult(result));
-
-        ObTableQueryAsyncResult obTableQueryAsyncResult = (ObTableQueryAsyncResult) result;
-        isEnd = obTableQueryAsyncResult.isEnd();
-        sessionId = obTableQueryAsyncResult.getSessionId();
-        return (ObTableQueryAsyncResult) result;
-    }
-
-    public boolean isEnd() {
-        return isEnd;
-    }
-
-    public void setEnd(boolean end) {
-        isEnd = end;
-    }
-
-    public long getSessionId() {
-        return sessionId;
-    }
-
-    public void setSessionId(long sessionId) {
-        this.sessionId = sessionId;
-    }
-}
