+++ src/main/java/com/alipay/oceanbase/rpc/table/api/Table.java
@@ -29,10 +29,6 @@ public interface Table {
 
     TableQuery query(String tableName) throws Exception;
 
-    TableQuery queryByBatchV2(String tableName) throws Exception;
-
-    TableQuery queryByBatch(String tableName) throws Exception;
-
     TableBatchOps batch(String tableName) throws Exception;
 
     Map<String, Object> get(String tableName, Object rowkey, String[] columns) throws Exception;
